+++ react/rules/ARG.js
@@ -1,5 +1,11 @@
+import argentina from '@vtex/phone/countries/ARG'
+import { getPhoneFields } from '../modules/phone'
+import regexValidation from '../modules/regexValidation'
+
+const phoneCountryCode = '54'
+
 export default {
-  country: 'UNI',
+  country: 'ARG',
   personalFields: [
     {
       name: 'firstName',
@@ -22,13 +28,15 @@ export default {
     {
       name: 'document',
       maxLength: 50,
-      label: 'document',
+      label: 'ARG_dni',
       required: true,
+      validate: regexValidation(/^[A-z]?\d{6,8}$/),
     },
     {
       name: 'homePhone',
       maxLength: 30,
       label: 'homePhone',
+      ...getPhoneFields(phoneCountryCode),
     },
     {
       name: 'gender',
@@ -55,12 +63,14 @@ export default {
     {
       name: 'corporateDocument',
       maxLength: 30,
-      label: 'corporateDocument',
+      label: 'ARG_cuit',
+      validate: regexValidation(/^([\d]{2})\-?([\d]{8})\-?([\d]{1})$/),
     },
     {
       name: 'businessPhone',
       maxLength: 30,
       label: 'businessPhone',
+      ...getPhoneFields(phoneCountryCode),
     },
   ],
 }