+++ react/rules/GBR.js
@@ -1,5 +1,11 @@
+import gbr from '@vtex/phone/countries/GBR'
+import { getPhoneFields } from '../modules/phone'
+import regexValidation from '../modules/regexValidation'
+
+const phoneCountryCode = '44'
+
 export default {
-  country: 'UNI',
+  country: 'GBR',
   personalFields: [
     {
       name: 'firstName',
@@ -20,15 +26,10 @@ export default {
       hidden: true,
     },
     {
-      name: 'document',
-      maxLength: 50,
-      label: 'document',
-      required: true,
-    },
-    {
       name: 'homePhone',
       maxLength: 30,
       label: 'homePhone',
+      ...getPhoneFields(phoneCountryCode),
     },
     {
       name: 'gender',
@@ -53,14 +54,10 @@ export default {
       label: 'tradeName',
     },
     {
-      name: 'corporateDocument',
-      maxLength: 30,
-      label: 'corporateDocument',
-    },
-    {
       name: 'businessPhone',
       maxLength: 30,
       label: 'businessPhone',
+      ...getPhoneFields(phoneCountryCode),
     },
   ],
 }